Fix new file path timing issue in `-setDocument:`
Sometime recently, TextMate stopped updating the window's representedFilename synchronously as part of `OakTextView -setDocument:`, which means we started getting the path for the file that the user was switching *away from* instead of the one they were switching to!

Fortunately, TextMate has also added a nice ObjC OakDocument class since this code was last worked on (used to be a C++ class; hard to dynamically inspect), so we can now simply ask the document being set for its path. Much simpler :)

A future update should come back around and clear out the old code here that supports TM1 and older versions of TM2.