Update Xcode and EditorConfig-Core + code signing
Makes the project compatible with Xcode 11 and adds code signing for the bundle we build. This is required to work with TextMate 2.x (non-beta versions). The new project configuration here is not really ideal -- it generates lots of warnings when building and uses an old format, but I want to get something working out the door without changing it too much. This already required a fair amount of weird finagling to get code signing to work right and to get PCRE2 (required for updated EditorConfig-Core) linking correctly.

This also updates EditorConfig-Core to v0.12.3 (the latest as of now).